refactor(front): move JobsStatusBadges view
Move this view in tests subdir, as its purpose is only to visually test
all possible values.
