+++ frontend/src/views/tests/JobsStatusBadges.vue
